Basicflaskapp/requirements.txt
boto3==1.7.33
botocore==1.10.33
certifi==2018.4.16
chardet==3.0.4
click==6.7
cloudant==2.4.0
cycler==0.10.0
docutils==0.14
Flask==0.12.2
Flask-DB2==0.0.10
ibm-cos-sdk==2.1.1
ibm-cos-sdk-core==2.1.1
ibm-cos-sdk-s3transfer==2.1.1
ibm-db==2.0.8a0
idna==2.6
itsdangerous==0.24
Jinja2==2.10
jmespath==0.9.3
kiwisolver==1.0.1
MarkupSafe==1.0
matplotlib==2.2.2
numpy==1.14.4
pandas==0.23.0
PyMySQL==0.8.1
pyparsing==2.2.0
python-dateutil==2.7.3
python-memcached==1.59
pytz==2018.4
requests==2.18.4
s3transfer==0.1.13
scikit-learn==0.19.1
scipy==1.1.0
six==1.11.0
sklearn==0.0
urllib3==1.22
Werkzeug==0.14.1